+++ JK.Common.Tests/Specifications/LongitudeSpecificationTests.cs
@@ -11,7 +11,7 @@ namespace JK.Common.Tests.Specifications
         [InlineData(0, true)]
         [InlineData(180, true)]
         [InlineData(181, false)]
-        public void IsSatisfiedBy(decimal input, bool expected)
+        public void IsSatisfiedBy(double input, bool expected)
         {
             var specification = new LongitudeSpecification();
             bool actual = specification.IsSatisfiedBy(input);
