+++ JK.Common/Specifications/LongitudeSpecification.cs
@@ -2,9 +2,9 @@
 
 namespace JK.Common.Specifications
 {
-    public class LongitudeSpecification : Specification<decimal>
+    public class LongitudeSpecification : Specification<double>
     {
-        public override bool IsSatisfiedBy(decimal candidate)
+        public override bool IsSatisfiedBy(double candidate)
         {
             return candidate >= -180 && candidate <= 180;
         }
